Match all multicast and broadcast in one flow.
Thanks to Richard Sanger for pointing this out to me:
Ethernet frames with a value of 1 in the least-significant bit of the first octet[note 2] of the destination address are treated as multicast frames and are flooded to all points on the network. While frames with ones in all bits of the destination address (FF-FF-FF-FF-FF-FF) are sometimes referred to as broadcasts, Ethernet network equipment generally does not distinguish between multicast and broadcast frames.

(cherry picked from commit 72f3f95cfbdda5e050464059b9baec32541c6c09)